% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Batch_1.R
\docType{data}
\name{Batch_1}
\alias{Batch_1}
\title{Batch_1 datasets for the examples.}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 500 rows and 30 columns.
}
\usage{
Batch_1
}
\description{
#These test datasets are part of the real gene expression datasets, including  three batches of samples.
}
\keyword{datasets}
