% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Batch_2.R
\docType{data}
\name{Batch_2}
\alias{Batch_2}
\title{Batch_2 datasets for the examples.}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 500 rows and 24 columns.
}
\usage{
Batch_2
}
\description{
These test datasets are part of the real gene expression datasets, including  three batches of samples.
}
\keyword{datasets}
