% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Batch_3.R
\docType{data}
\name{Batch_3}
\alias{Batch_3}
\title{Batch_3 datasets for the examples.}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 500 rows and 20 columns.
}
\usage{
Batch_3
}
\description{
#These test datasets are part of the real gene expression datasets, including  three batches of samples.
}
\examples{
data(Test_data)
}
\keyword{datasets}
