% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Group_1.R
\docType{data}
\name{Group_1}
\alias{Group_1}
\title{Group of Batch_1 datasets.}
\format{
An object of class \code{character} of length 30.
}
\usage{
Group_1
}
\description{
# The sample_A_B represents the B sample in batch A. tumor/normal  represent tumor group and normal group respectively.
}
\examples{
data(Test_data)
}
\keyword{datasets}
