% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Group_2.R
\docType{data}
\name{Group_2}
\alias{Group_2}
\title{Group of Batch_2 datasets.}
\format{
An object of class \code{character} of length 24.
}
\usage{
Group_2
}
\description{
# The sample_A_B represents the B sample in batch A. tumor/normal  represent tumor group and normal group respectively.
}
\keyword{datasets}
