% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Group_3.R
\docType{data}
\name{Group_3}
\alias{Group_3}
\title{Group of Batch_3 datasets.}
\format{
An object of class \code{character} of length 20.
}
\usage{
Group_3
}
\description{
# The sample_A_B represents the B sample in batch A. tumor/normal  represent tumor group and normal group respectively.
}
\keyword{datasets}
